perf: split job management
