+++ src/views/ops/Template/Adhoc.vue
@@ -49,7 +49,3 @@ export default {
   }
 }
 </script>
-
-<style>
-
-</style>
