+++ src/views/ops/Template/index.vue
@@ -36,12 +36,6 @@ export default {
         }
       ]
     }
-  },
-  mounted() {
   }
 }
 </script>
-
-<style scoped>
-
-</style>
