+++ src/views/sessions/SessionList/BaseList.vue
@@ -146,8 +146,7 @@ export default {
         }
       }
     }
-  },
-  methods: {}
+  }
 }
 </script>
